Create AbstractFilteredProvider for sortable filtered providers
